+++ plugins/modules/fortiflexvm_configs_create.py
@@ -32,6 +32,10 @@ options:
             - The password to authenticate. If not declared, the code will read the environment variable FORTIFLEX_ACCESS_PASSWORD.
         type: str
         required: false
+    accountId:
+        description: Account ID.
+        type: int
+        required: false
     programSerialNumber:
         description:
             - The serial number of your Flex VM Program.
@@ -149,6 +153,41 @@ options:
                 elements: str
                 required: false
                 default: []
+    fortiClientEMSOP:
+        description:
+            - FortiClient EMS On-Prem.
+        type: dict
+        required: false
+        suboptions:
+            ZTNA:
+                description:
+                    - ZTNA/VPN (number of endpoints).
+                    - Number between 0 and 25000 (inclusive). Value should be divisible by 25.
+                type: int
+                required: true
+            EPP:
+                description:
+                    - EPP/ATP + ZTNA/VPN (number of endpoints).
+                    - Number between 0 and 25000 (inclusive). Value should be divisible by 25.
+                type: int
+                required: true
+            chromebook:
+                description:
+                    - Chromebook (number of endpoints).
+                    - Number between 0 and 25000 (inclusive). Value should be divisible by 25.
+                type: int
+                required: true
+            service:
+                description:
+                    - Support Services. Possible value is "FCTFC247" (FortiCare Premium)
+                type: str
+                required: true
+            addons:
+                description: Addons. A list. Possible value is "BPS" ( FortiCare Best Practice).
+                type: list
+                elements: str
+                required: false
+                default: []
     fortiAnalyzer:
         description:
             - FortiAnalyzer Virtual Machine.
@@ -211,7 +250,10 @@ options:
                     - FGT40F (FortiGate-40F), FGT60F (FortiGate-60F), FGT70F (FortiGate-70F), FGT80F (FortiGate-80F),
                     - FG100F (FortiGate-100F), FGT60E (FortiGate-60E), FGT61F (FortiGate-61F), FG100E (FortiGate-100E),
                     - FG101F (FortiGate-101F), FG200E (FortiGate-200E), FG200F (FortiGate-200F), FG201F (FortiGate-201F),
-                    - FG4H0F (FortiGate-400F), FG6H0F (FortiGate-600F).
+                    - FG4H0F (FortiGate-400F), FG6H0F (FortiGate-600F), FWF40F (FortiWifi-40F), FWF60F (FortiWifi-60F),
+                    - FGR60F (FortiGateRugged-60F), FR70FB (FortiGateRugged-70F), FGT81F (FortiGate-81F), FG101E (FortiGate-101E),
+                    - FG4H1F (FortiGate-401F), FG1K0F (FortiGate-1000F), FG180F (FortiGate-1800F), F2K60F (FortiGate-2600F),
+                    - FG3K0F (FortiGate-3000F), FG3K1F (FortiGate-3001F), FG3K2F (FortiGate-3200F).
                 type: str
                 required: true
             service:
@@ -222,10 +264,53 @@ options:
                 required: true
             addons:
                 description:
-                    - Addons. Only support "NONE" now, will support "FGHWFCELU" (FortiCare Elite Upgrade) in the future.
-                type: str
+                    - Addons. A list, can be empty, possible values are
+                    - FGHWFCELU (FortiCare Elite Upgrade), FGHWFAMS (FortiGate Cloud Management),
+                    - FGHWFAIS (AI-Based In-line Sandbox), FGHWSWNM (SD-WAN Underlay), FGHWDLDB (FortiGuard DLP),
+                    - FGHWFAZC (FortiAnalyzer Cloud), FGHWSOCA (SOCaaS), FGHWMGAS (Managed FortiGate),
+                    - FGHWSPAL (SD-WAN Connector for FortiSASE), FGHWFCSS (FortiConverter Service).
+                type: list
+                elements: str
                 required: false
-                default: "NONE"
+                default: []
+    fortiCloudPrivate:
+        description:
+            - FortiWeb Cloud, Private.
+        type: dict
+        required: false
+        version_added: 2.0.0
+        suboptions:
+            throughput:
+                description:
+                    - Average Throughput (Mbps).
+                    - Possible values are 10, 25, 50, 75, 100, 150, 200, 250, 300, 350, 400, 450, 500, 600,
+                    - 700, 800, 900, 1000, 1500, 2000, 2500, 3000, 3500, 4000, 4500, 5000, 5500, 6000, 6500,
+                    - 7000, 7500, 8000, 8500, 9000, 9500, 10000.
+                type: int
+                required: true
+            applications:
+                description: Number of web applications. Number between 0 and 2000 (inclusive).
+                type: int
+                required: true
+    fortiCloudPublic:
+        description:
+            - FortiWeb Cloud, Public.
+        type: dict
+        required: false
+        version_added: 2.0.0
+        suboptions:
+            throughput:
+                description:
+                    - Average Throughput (Mbps).
+                    - Possible values are 10, 25, 50, 75, 100, 150, 200, 250, 300, 350, 400, 450, 500, 600,
+                    - 700, 800, 900, 1000, 1500, 2000, 2500, 3000, 3500, 4000, 4500, 5000, 5500, 6000, 6500,
+                    - 7000, 7500, 8000, 8500, 9000, 9500, 10000.
+                type: int
+                required: true
+            applications:
+                description: Number of web applications. Number between 0 and 2000 (inclusive).
+                type: int
+                required: true
 '''
 
 EXAMPLES = '''
@@ -256,9 +341,9 @@ EXAMPLES = '''
 
         # Please only use one of the following.
         fortiGateBundle:
-          cpu: "2"      # "1", "2", "4", "8", "16", "32", "2147483647"
-          service: "FC" # "FC", "UTP", "ENT", "ATP"
-          vdom: 10      # 0 ~ 500
+          cpu: "2"                            # "1", "2", "4", "8", "16", "32", "2147483647"
+          service: "UTP"                      # "FC", "UTP", "ENT", "ATP"
+          vdom: 10                            # 0 ~ 500
 
         # fortiManager:
         #   device: 1                         # 1 ~ 100000
@@ -275,6 +360,13 @@ EXAMPLES = '''
         #   vdom: 1                           # 1 ~ 500
         #   cloudServices: ["FAMS", "SWNM"]   # "FAMS", "SWNM", "AFAC", "FAZC"
 
+        # fortiClientEMSOP:
+        #   ZTNA: 1000                        # 0 ~ 25000. Value should be divisible by 25.
+        #   EPP: 1000                         # 0 ~ 25000. Value should be divisible by 25.
+        #   chromebook: 1000                  # 0 ~ 25000. Value should be divisible by 25.
+        #   service: "FCTFC247"               # "FCTFC247"
+        #   addons: ["BPS"]                   # Empty or "BPS"
+
         # fortiAnalyzer:
         #   storage: 5                        # 5 ~ 8300
         #   adom: 1                           # 0 ~ 1200
@@ -284,14 +376,29 @@ EXAMPLES = '''
         #   device: 1                         # 0 ~ 100000
 
         # fortiADC:
-        #   cpu: "1"                          # "1", "2", "4", "8", "16", "32"
+        #   cpu: "32"                         # "1", "2", "4", "8", "16", "32"
         #   service: "FDVSTD"                 # "FDVSTD", "FDVADV" or "FDVFC247"
 
         # fortiGateHardware:
-        #   model: "FGT40F"                   # "FGT40F", "FGT60F", "FGT70F", "FGT80F", "FG100F", "FGT60E", "FGT61F",
-        #                                     # "FG100E", "FG101F", "FG200E", "FG200F", "FG201F", "FG4H0F", "FG6H0F"
-        #   service: "FGHWFC247"              # "FGHWFC247", "FGHWFCEL", "FDVFC247", "FGHWUTP" or "FGHWENT"
-        #   addons: "NONE"
+        #   model: "FGT60F"                   # "FGT40F", "FGT60F", "FGT70F", "FGT80F", "FG100F", "FGT60E", "FGT61F",
+        #                                     # "FG100E", "FG101F", "FG200E", "FG200F", "FG201F", "FG4H0F", "FG6H0F",
+        #                                     # "FWF40F", "FWF60F", "FGR60F", "FR70FB", "FGT81F", "FG101E", "FG4H1F",
+        #                                     # "FG1K0F", "FG180F", "F2K60F", "FG3K0F", "FG3K1F", "FG3K2F"
+        #   service: "FGHWFCEL"               # "FGHWFC247", "FGHWFCEL", "FDVFC247", "FGHWUTP" or "FGHWENT"
+        #   addons: []                        # "FGHWFCELU", "FGHWFAMS", "FGHWFAIS", "FGHWSWNM", "FGHWDLDB",
+        #                                     # "FGHWFAZC", "FGHWSOCA", "FGHWMGAS", "FGHWSPAL", "FGHWFCSS"
+
+        # fortiCloudPrivate:
+        #   throughput: 100                   # 10, 25, 50, 75, 100, 150, 200, 250, 300, 350, 400, 450, 500, 600, 700, 800,
+        #                                     # 900, 1000, 1500, 2000, 2500, 3000, 3500, 4000, 4500, 5000, 5500, 6000, 6500,
+        #                                     # 7000, 7500, 8000, 8500, 9000, 9500, 10000.
+        #   applications: 10                  # 0 ~ 2000
+
+        # fortiCloudPublic:
+        #   throughput: 100                   # 10, 25, 50, 75, 100, 150, 200, 250, 300, 350, 400, 450, 500, 600, 700, 800,
+        #                                     # 900, 1000, 1500, 2000, 2500, 3000, 3500, 4000, 4500, 5000, 5500, 6000, 6500,
+        #                                     # 7000, 7500, 8000, 8500, 9000, 9500, 10000.
+        #   applications: 10                  # 0 ~ 2000
 
       register: result
 
@@ -306,6 +413,11 @@ configs:
     type: dict
     returned: always
     contains:
+        accountId:
+            description: The ID of the account associated with the program.
+            type: int
+            returned: always
+            sample: 12345
         id:
             description: The ID of the configuration.
             type: int
@@ -330,7 +442,6 @@ configs:
             description:
                 - FortiGate Virtual Machine - Service Bundle.
             type: dict
-            returned: changed
             contains:
                 cpu:
                     description:
@@ -339,7 +450,7 @@ configs:
                     returned: always
                 service:
                     description:
-                        - he value of this attribute is one of "FC" (FortiCare), "UTP", "ENT" (Enterprise) or "ATP".
+                        - The value of this attribute is one of "FC" (FortiCare), "UTP", "ENT" (Enterprise) or "ATP".
                     type: str
                     returned: always
                 vdom:
@@ -351,7 +462,6 @@ configs:
             description:
                 - FortiManager Virtual Machine.
             type: dict
-            returned: changed
             contains:
                 device:
                     description:
@@ -367,7 +477,6 @@ configs:
             description:
                 - FortiWeb Virtual Machine - Service Bundle.
             type: dict
-            returned: changed
             contains:
                 cpu:
                     description:
@@ -383,7 +492,6 @@ configs:
             description:
                 - FortiGate Virtual Machine - A La Carte Services.
             type: dict
-            returned: changed
             contains:
                 cpu:
                     description:
@@ -395,6 +503,7 @@ configs:
                         - The fortiguard services this FortiGate Virtual Machine supports. The default value is an empty list.
                         - It should contain zero, one or more elements of ["IPS", "AVDB", "FGSA", "DLDB", "FAIS", "FURLDNS"].
                     type: list
+                    elements: str
                     returned: always
                 supportService:
                     description:
@@ -411,12 +520,44 @@ configs:
                         - The cloud services this FortiGate Virtual Machine supports. The default value is an empty list.
                         - It should contain zero, one or more elements of ["FAMS", "SWNM", "AFAC", "FAZC"].
                     type: list
+                    elements: str
                     returned: always
+        fortiClientEMSOP:
+            description:
+                - FortiClient EMS On-Prem.
+            type: dict
+            contains:
+                ZTNA:
+                    description:
+                        - ZTNA/VPN (number of endpoints).
+                        - Number between 0 and 25000 (inclusive). Value should be divisible by 25.
+                    type: int
+                    returned: always
+                EPP:
+                    description:
+                        - EPP/ATP + ZTNA/VPN (number of endpoints).
+                        - Number between 0 and 25000 (inclusive). Value should be divisible by 25.
+                    type: int
+                    returned: always
+                chromebook:
+                    description:
+                        - Chromebook (number of endpoints).
+                        - Number between 0 and 25000 (inclusive). Value should be divisible by 25.
+                    type: int
+                    returned: always
+                service:
+                    description:
+                        - Support Services. Possible value is "FCTFC247" (FortiCare Premium)
+                    type: str
+                    returned: always
+                addons:
+                    description: Addons. A list. Possible value is "BPS" ( FortiCare Best Practice).
+                    type: list
+                    elements: str
         fortiAnalyzer:
             description:
                 - FortiAnalyzer Virtual Machine.
             type: dict
-            returned: changed
             contains:
                 storage:
                     description:
@@ -438,7 +579,6 @@ configs:
             description:
                 - FortiPortal Virtual Machine.
             type: dict
-            returned: changed
             contains:
                 device:
                     description:
@@ -449,7 +589,6 @@ configs:
             description:
                 - FortiADC Virtual Machine.
             type: dict
-            returned: changed
             contains:
                 cpu:
                     description:
@@ -465,7 +604,6 @@ configs:
             description:
                 - FortiGate Hardware.
             type: dict
-            returned: changed
             contains:
                 model:
                     description:
@@ -473,7 +611,10 @@ configs:
                         - FGT40F (FortiGate-40F), FGT60F (FortiGate-60F), FGT70F (FortiGate-70F), FGT80F (FortiGate-80F),
                         - FG100F (FortiGate-100F), FGT60E (FortiGate-60E), FGT61F (FortiGate-61F), FG100E (FortiGate-100E),
                         - FG101F (FortiGate-101F), FG200E (FortiGate-200E), FG200F (FortiGate-200F), FG201F (FortiGate-201F),
-                        - FG4H0F (FortiGate-400F), FG6H0F (FortiGate-600F).
+                        - FG4H0F (FortiGate-400F), FG6H0F (FortiGate-600F), FWF40F (FortiWifi-40F), FWF60F (FortiWifi-60F),
+                        - FGR60F (FortiGateRugged-60F), FR70FB (FortiGateRugged-70F), FGT81F (FortiGate-81F), FG101E (FortiGate-101E),
+                        - FG4H1F (FortiGate-401F), FG1K0F (FortiGate-1000F), FG180F (FortiGate-1800F), F2K60F (FortiGate-2600F),
+                        - FG3K0F (FortiGate-3000F), FG3K1F (FortiGate-3001F), FG3K2F (FortiGate-3200F).
                     type: str
                     returned: always
                 service:
@@ -484,8 +625,47 @@ configs:
                     returned: always
                 addons:
                     description:
-                        - Addons. Only support "NONE" now, will support "FGHWFCELU" (FortiCare Elite Upgrade) in the future.
-                    type: str
+                        - Addons. Possible values are
+                        - NONE, FGHWFCELU (FortiCare Elite Upgrade), FGHWFAMS (FortiGate Cloud Management),
+                        - FGHWFAIS (AI-Based In-line Sandbox), FGHWSWNM (SD-WAN Underlay), FGHWDLDB (FortiGuard DLP),
+                        - FGHWFAZC (FortiAnalyzer Cloud), FGHWSOCA (SOCaaS), FGHWMGAS (Managed FortiGate),
+                        - FGHWSPAL (SD-WAN Connector for FortiSASE), FGHWFCSS (FortiConverter Service).
+                    type: list
+                    elements: str
+                    returned: always
+        fortiCloudPrivate:
+            description:
+                - FortiWeb Cloud, Private.
+            type: dict
+            contains:
+                throughput:
+                    description:
+                        - Average Throughput (Mbps).
+                        - Possible values are 10, 25, 50, 75, 100, 150, 200, 250, 300, 350, 400, 450, 500, 600,
+                        - 700, 800, 900, 1000, 1500, 2000, 2500, 3000, 3500, 4000, 4500, 5000, 5500, 6000, 6500,
+                        - 7000, 7500, 8000, 8500, 9000, 9500, 10000.
+                    type: int
+                    returned: always
+                applications:
+                    description: Number of web applications. Number between 0 and 2000 (inclusive).
+                    type: int
+                    returned: always
+        fortiCloudPublic:
+            description:
+                - FortiWeb Cloud, Public.
+            type: dict
+            contains:
+                throughput:
+                    description:
+                        - Average Throughput (Mbps).
+                        - Possible values are 10, 25, 50, 75, 100, 150, 200, 250, 300, 350, 400, 450, 500, 600,
+                        - 700, 800, 900, 1000, 1500, 2000, 2500, 3000, 3500, 4000, 4500, 5000, 5500, 6000, 6500,
+                        - 7000, 7500, 8000, 8500, 9000, 9500, 10000.
+                    type: int
+                    returned: always
+                applications:
+                    description: Number of web applications. Number between 0 and 2000 (inclusive).
+                    type: int
                     returned: always
 '''
 
@@ -499,6 +679,7 @@ def get_module_args():
     module_args = dict(
         username=dict(type="str", required=False),
         password=dict(type="str", required=False, no_log=True),
+        accountId=dict(type='int', required=False),
         programSerialNumber=dict(type="str", required=True),
         name=dict(type="str", required=True),
         bypass_validation=dict(type="bool", required=False, default=False),
@@ -545,6 +726,8 @@ def main():
         "productTypeId": product_id,
         "parameters": parameters
     }
+    if module.params["accountId"]:
+        data["accountId"] = module.params["accountId"]
 
     # Check mode
     if module.check_mode: