+++ plugins/modules/fortiflexvm_configs_list_info.py
@@ -32,6 +32,10 @@ options:
             - The password to authenticate. If not declared, the code will read the environment variable FORTIFLEX_ACCESS_PASSWORD.
         type: str
         required: false
+    accountId:
+        description: Account ID.
+        type: int
+        required: false
     programSerialNumber:
         description:
             - The serial number of the program to get configs for.
@@ -52,6 +56,7 @@ EXAMPLES = '''
       fortinet.fortiflexvm.fortiflexvm_configs_list_info:
         username: "{{ username }}"
         password: "{{ password }}"
+        # accountId: 12345 # optional
         programSerialNumber: "ELAVMS000000XXXX"
       register: result
 
@@ -66,6 +71,11 @@ configs:
     type: list
     returned: always
     contains:
+        accountId:
+            description: The ID of the account associated with the program.
+            type: int
+            returned: always
+            sample: 12345
         id:
             description: The ID of the configuration.
             type: int
@@ -90,7 +100,6 @@ configs:
             description:
                 - FortiGate Virtual Machine - Service Bundle.
             type: dict
-            returned: changed
             contains:
                 cpu:
                     description:
@@ -99,7 +108,7 @@ configs:
                     returned: always
                 service:
                     description:
-                        - he value of this attribute is one of "FC" (FortiCare), "UTM", "ENT" (Enterprise) or "ATP".
+                        - The value of this attribute is one of "FC" (FortiCare), "UTP", "ENT" (Enterprise) or "ATP".
                     type: str
                     returned: always
                 vdom:
@@ -111,7 +120,6 @@ configs:
             description:
                 - FortiManager Virtual Machine.
             type: dict
-            returned: changed
             contains:
                 device:
                     description:
@@ -127,7 +135,6 @@ configs:
             description:
                 - FortiWeb Virtual Machine - Service Bundle.
             type: dict
-            returned: changed
             contains:
                 cpu:
                     description:
@@ -143,7 +150,6 @@ configs:
             description:
                 - FortiGate Virtual Machine - A La Carte Services.
             type: dict
-            returned: changed
             contains:
                 cpu:
                     description:
@@ -155,6 +161,7 @@ configs:
                         - The fortiguard services this FortiGate Virtual Machine supports. The default value is an empty list.
                         - It should contain zero, one or more elements of ["IPS", "AVDB", "FGSA", "DLDB", "FAIS", "FURLDNS"].
                     type: list
+                    elements: str
                     returned: always
                 supportService:
                     description:
@@ -171,12 +178,44 @@ configs:
                         - The cloud services this FortiGate Virtual Machine supports. The default value is an empty list.
                         - It should contain zero, one or more elements of ["FAMS", "SWNM", "AFAC", "FAZC"].
                     type: list
+                    elements: str
                     returned: always
+        fortiClientEMSOP:
+            description:
+                - FortiClient EMS On-Prem.
+            type: dict
+            contains:
+                ZTNA:
+                    description:
+                        - ZTNA/VPN (number of endpoints).
+                        - Number between 0 and 25000 (inclusive). Value should be divisible by 25.
+                    type: int
+                    returned: always
+                EPP:
+                    description:
+                        - EPP/ATP + ZTNA/VPN (number of endpoints).
+                        - Number between 0 and 25000 (inclusive). Value should be divisible by 25.
+                    type: int
+                    returned: always
+                chromebook:
+                    description:
+                        - Chromebook (number of endpoints).
+                        - Number between 0 and 25000 (inclusive). Value should be divisible by 25.
+                    type: int
+                    returned: always
+                service:
+                    description:
+                        - Support Services. Possible value is "FCTFC247" (FortiCare Premium)
+                    type: str
+                    returned: always
+                addons:
+                    description: Addons. A list. Possible value is "BPS" ( FortiCare Best Practice).
+                    type: list
+                    elements: str
         fortiAnalyzer:
             description:
                 - FortiAnalyzer Virtual Machine.
             type: dict
-            returned: changed
             contains:
                 storage:
                     description:
@@ -198,7 +237,6 @@ configs:
             description:
                 - FortiPortal Virtual Machine.
             type: dict
-            returned: changed
             contains:
                 device:
                     description:
@@ -209,7 +247,6 @@ configs:
             description:
                 - FortiADC Virtual Machine.
             type: dict
-            returned: changed
             contains:
                 cpu:
                     description:
@@ -225,7 +262,6 @@ configs:
             description:
                 - FortiGate Hardware.
             type: dict
-            returned: changed
             contains:
                 model:
                     description:
@@ -233,7 +269,10 @@ configs:
                         - FGT40F (FortiGate-40F), FGT60F (FortiGate-60F), FGT70F (FortiGate-70F), FGT80F (FortiGate-80F),
                         - FG100F (FortiGate-100F), FGT60E (FortiGate-60E), FGT61F (FortiGate-61F), FG100E (FortiGate-100E),
                         - FG101F (FortiGate-101F), FG200E (FortiGate-200E), FG200F (FortiGate-200F), FG201F (FortiGate-201F),
-                        - FG4H0F (FortiGate-400F), FG6H0F (FortiGate-600F).
+                        - FG4H0F (FortiGate-400F), FG6H0F (FortiGate-600F), FWF40F (FortiWifi-40F), FWF60F (FortiWifi-60F),
+                        - FGR60F (FortiGateRugged-60F), FR70FB (FortiGateRugged-70F), FGT81F (FortiGate-81F), FG101E (FortiGate-101E),
+                        - FG4H1F (FortiGate-401F), FG1K0F (FortiGate-1000F), FG180F (FortiGate-1800F), F2K60F (FortiGate-2600F),
+                        - FG3K0F (FortiGate-3000F), FG3K1F (FortiGate-3001F), FG3K2F (FortiGate-3200F).
                     type: str
                     returned: always
                 service:
@@ -244,8 +283,47 @@ configs:
                     returned: always
                 addons:
                     description:
-                        - Addons. Only support "NONE" now, will support "FGHWFCELU" (FortiCare Elite Upgrade) in the future.
-                    type: str
+                        - Addons. A list, possible values are
+                        - FGHWFCELU (FortiCare Elite Upgrade), FGHWFAMS (FortiGate Cloud Management),
+                        - FGHWFAIS (AI-Based In-line Sandbox), FGHWSWNM (SD-WAN Underlay), FGHWDLDB (FortiGuard DLP),
+                        - FGHWFAZC (FortiAnalyzer Cloud), FGHWSOCA (SOCaaS), FGHWMGAS (Managed FortiGate),
+                        - FGHWSPAL (SD-WAN Connector for FortiSASE), FGHWFCSS (FortiConverter Service).
+                    type: list
+                    elements: str
+                    returned: always
+        fortiCloudPrivate:
+            description:
+                - FortiWeb Cloud, Private.
+            type: dict
+            contains:
+                throughput:
+                    description:
+                        - Average Throughput (Mbps).
+                        - Possible values are 10, 25, 50, 75, 100, 150, 200, 250, 300, 350, 400, 450, 500, 600,
+                        - 700, 800, 900, 1000, 1500, 2000, 2500, 3000, 3500, 4000, 4500, 5000, 5500, 6000, 6500,
+                        - 7000, 7500, 8000, 8500, 9000, 9500, 10000.
+                    type: int
+                    returned: always
+                applications:
+                    description: Number of web applications. Number between 0 and 2000 (inclusive).
+                    type: int
+                    returned: always
+        fortiCloudPublic:
+            description:
+                - FortiWeb Cloud, Public.
+            type: dict
+            contains:
+                throughput:
+                    description:
+                        - Average Throughput (Mbps).
+                        - Possible values are 10, 25, 50, 75, 100, 150, 200, 250, 300, 350, 400, 450, 500, 600,
+                        - 700, 800, 900, 1000, 1500, 2000, 2500, 3000, 3500, 4000, 4500, 5000, 5500, 6000, 6500,
+                        - 7000, 7500, 8000, 8500, 9000, 9500, 10000.
+                    type: int
+                    returned: always
+                applications:
+                    description: Number of web applications. Number between 0 and 2000 (inclusive).
+                    type: int
                     returned: always
 '''
 
@@ -259,6 +337,7 @@ def main():
     module_args = dict(
         username=dict(type="str", required=False),
         password=dict(type="str", required=False, no_log=True),
+        accountId=dict(type='int', required=False),
         programSerialNumber=dict(type="str", required=True),
     )
 
@@ -272,7 +351,9 @@ def main():
     connection = Connection(module, module.params["username"], module.params["password"])
 
     # Send request to get config list
-    data = {"programSerialNumber": module.params['programSerialNumber']}
+    data = {"programSerialNumber": module.params["programSerialNumber"]}
+    if module.params["accountId"]:
+        data["accountId"] = module.params["accountId"]
     response = connection.send_request("fortiflex/v2/configs/list", data, method="POST")
 
     # Trasform the format of output data