+++ plugins/modules/fortiflexvm_entitlements_update.py
@@ -94,31 +94,36 @@ entitlements:
     type: list
     returned: always
     contains:
-        serialNumber:
-            description: The serial number of the entitlement.
-            type: str
+        accountId:
+            description: The ID of the account associated with the program.
+            type: int
             returned: always
-            sample: "FGVMMLTM23001324"
-        description:
-            description: The description of the entitlement.
-            type: str
-            returned: always
-            sample: "Modify through Ansible"
+            sample: 12345
         configId:
             description: The config ID of the entitlement.
             type: int
             returned: always
             sample: 3196
-        startDate:
-            description: The start date of the entitlement.
+        description:
+            description: The description of the entitlement.
             type: str
             returned: always
-            sample: "2023-03-13T11:48:53.03"
+            sample: "Modify through Ansible"
         endDate:
             description: The end date of the entitlement.
             type: str
             returned: always
             sample: "2023-12-12T00:00:00"
+        serialNumber:
+            description: The serial number of the entitlement.
+            type: str
+            returned: always
+            sample: "FGVMMLTM23001324"
+        startDate:
+            description: The start date of the entitlement.
+            type: str
+            returned: always
+            sample: "2023-03-13T11:48:53.03"
         status:
             description: The status of the VM. Possible values are "PENDING", "ACTIVE", "STOPPED" or "EXPIRED".
             type: str